+++ tests/end2end/features/commandline/test_commandline_bdd.py
@@ -6,6 +6,7 @@
 
 import contextlib
 
+import pytest
 import pytest_bdd as bdd
 
 import vimiv
@@ -16,9 +17,22 @@ from vimiv.commands import runners
 bdd.scenarios("commandline.feature")
 
 
+@pytest.fixture()
+def history(commandline):
+    """Fixture to retrieve the commandline history object."""
+    return commandline._history
+
+
 @bdd.when("I run help <topic>")
 def run_help(topic):
     runners.run(f"help {topic}", mode=api.modes.current())
+
+
+@bdd.when("I populate the history")
+def populate_history(history):
+    commands = [f":random-command{i:02d}" for i in range(20)]
+    for historydeque in history.values():
+        historydeque.extend(commands)
 
 
 @bdd.then("the mode should not be command")
@@ -45,3 +59,19 @@ def check_help(message_widget, topic):
         assert setting.desc in message_widget.text()
         return
     raise AssertionError(f"Topic '{topic}' not found")
+
+
+@bdd.then("the history of all modes should be empty")
+def check_complete_history_empty(history):
+    for historydeque in history.values():
+        assert not historydeque
+
+
+@bdd.then(bdd.parsers.parse("the history of {modename} mode should be empty"))
+def check_mode_history_empty(history, modename):
+    assert not history[api.modes.get_by_name(modename)]
+
+
+@bdd.then(bdd.parsers.parse("the history of {modename} mode should not be empty"))
+def check_mode_history_not_empty(history, modename):
+    assert history[api.modes.get_by_name(modename)]